envoy/xds: Refactor existing code
Extract the XDSServer interface into a separate package
(pkg/envoy/xds/xds_server.go) to decouple consumers from the
concrete implementation. Move shared envoy data types into a new
model.go and socket utility helpers into pkg/envoy/util/sockets.go.

Update the CiliumEnvoyConfig parser, secret sync, proxy wiring,
and daemon command to use the new interface and package layout.

This refactoring prepares the codebase for adding an alternative
ADS (Aggregated Discovery Service) server implementation behind
the same XDSServer interface.
