+++ pkg/ciliumenvoyconfig/cec_resource_parser_test.go
@@ -7,8 +7,10 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
+	"maps"
 	"testing"
 
+	util "github.com/cilium/cilium/pkg/envoy/util"
 	"github.com/cilium/hive/hivetest"
 	cilium "github.com/cilium/proxy/go/cilium/api"
 	envoy_config_cluster "github.com/envoyproxy/go-control-plane/envoy/config/cluster/v3"
@@ -27,7 +29,7 @@ import (
 
 	"github.com/cilium/cilium/pkg/annotation"
 	"github.com/cilium/cilium/pkg/bpf"
-	"github.com/cilium/cilium/pkg/envoy"
+	"github.com/cilium/cilium/pkg/envoy/xds"
 	cilium_v2 "github.com/cilium/cilium/pkg/k8s/apis/cilium.io/v2"
 	"github.com/cilium/cilium/pkg/lock"
 )
@@ -108,7 +110,7 @@ func TestUpstreamInject(t *testing.T) {
 	// Empty options
 	//
 	var opts envoy_upstreams_http_v3.HttpProtocolOptions
-	changed, err := injectCiliumUpstreamL7Filter(envoy.GetAccessLogSocketPath(), &opts, false)
+	changed, err := injectCiliumUpstreamL7Filter(util.GetAccessLogSocketPath(), &opts, false)
 	assert.NoError(t, err)
 	assert.True(t, changed)
 	assert.NotNil(t, opts.HttpFilters)
@@ -123,7 +125,7 @@ func TestUpstreamInject(t *testing.T) {
 	assert.NotNil(t, opts.GetUseDownstreamProtocolConfig()) // no ALPN support
 
 	// already present
-	changed, err = injectCiliumUpstreamL7Filter(envoy.GetAccessLogSocketPath(), &opts, true)
+	changed, err = injectCiliumUpstreamL7Filter(util.GetAccessLogSocketPath(), &opts, true)
 	assert.NoError(t, err)
 	assert.False(t, changed)
 	assert.NotNil(t, opts.HttpFilters)
@@ -148,7 +150,7 @@ func TestUpstreamInject(t *testing.T) {
 			},
 		},
 	}
-	changed, err = injectCiliumUpstreamL7Filter(envoy.GetAccessLogSocketPath(), &opts, true)
+	changed, err = injectCiliumUpstreamL7Filter(util.GetAccessLogSocketPath(), &opts, true)
 	assert.NoError(t, err)
 	assert.True(t, changed)
 	assert.NotNil(t, opts.HttpFilters)
@@ -170,7 +172,7 @@ func TestUpstreamInject(t *testing.T) {
 			},
 		},
 	}
-	changed, err = injectCiliumUpstreamL7Filter(envoy.GetAccessLogSocketPath(), &opts, true)
+	changed, err = injectCiliumUpstreamL7Filter(util.GetAccessLogSocketPath(), &opts, true)
 	assert.NoError(t, err)
 	assert.True(t, changed)
 	assert.NotNil(t, opts.HttpFilters)
@@ -199,7 +201,7 @@ func TestUpstreamInject(t *testing.T) {
 			},
 		},
 	}
-	changed, err = injectCiliumUpstreamL7Filter(envoy.GetAccessLogSocketPath(), &opts, true)
+	changed, err = injectCiliumUpstreamL7Filter(util.GetAccessLogSocketPath(), &opts, true)
 	assert.Error(t, err)
 	assert.False(t, changed)
 	assert.ErrorContains(t, err, "filter after codec filter: name:\"cilium.l7policy\"")
@@ -316,10 +318,11 @@ func TestCiliumEnvoyConfig(t *testing.T) {
 	resources, err := parser.ParseResources("namespace", "name", cec.Spec.Resources, false, false, false, true)
 	require.NoError(t, err)
 	assert.Len(t, resources.Listeners, 1)
-	assert.Equal(t, "namespace/name/envoy-prometheus-metrics-listener", resources.Listeners[0].Name)
-	assert.Equal(t, uint32(10000), resources.Listeners[0].Address.GetSocketAddress().GetPortValue())
-	assert.Len(t, resources.Listeners[0].FilterChains, 1)
-	chain := resources.Listeners[0].FilterChains[0]
+	listener := resources.Listeners["namespace/name/envoy-prometheus-metrics-listener"]
+	assert.Equal(t, "namespace/name/envoy-prometheus-metrics-listener", listener.Name)
+	assert.Equal(t, uint32(10000), listener.Address.GetSocketAddress().GetPortValue())
+	assert.Len(t, listener.FilterChains, 1)
+	chain := listener.FilterChains[0]
 
 	assert.NotNil(t, chain.TransportSocket)
 	assert.Equal(t, "envoy.transport_sockets.tls", chain.TransportSocket.Name)
@@ -371,7 +374,8 @@ func TestCiliumEnvoyConfig(t *testing.T) {
 	//
 	// Check that secret name was qualified
 	//
-	assert.Equal(t, "namespace/name/validation_context", resources.Secrets[0].Name)
+	secret := resources.Secrets["namespace/name/validation_context"]
+	assert.Equal(t, "namespace/name/validation_context", secret.Name)
 }
 
 var ciliumEnvoyConfigInvalid = `apiVersion: cilium.io/v2
@@ -419,10 +423,11 @@ func TestCiliumEnvoyConfigValidation(t *testing.T) {
 	resources, err := parser.ParseResources("namespace", "name", cec.Spec.Resources, false, false, false, false)
 	require.NoError(t, err)
 	assert.Len(t, resources.Listeners, 1)
-	assert.Equal(t, "namespace/name/envoy-prometheus-metrics-listener", resources.Listeners[0].Name)
-	assert.Equal(t, uint32(0), resources.Listeners[0].Address.GetSocketAddress().GetPortValue()) // invalid listener port number
-	assert.Len(t, resources.Listeners[0].FilterChains, 1)
-	chain := resources.Listeners[0].FilterChains[0]
+	listener := resources.Listeners["namespace/name/envoy-prometheus-metrics-listener"]
+	assert.Equal(t, "namespace/name/envoy-prometheus-metrics-listener", listener.Name)
+	assert.Equal(t, uint32(0), listener.Address.GetSocketAddress().GetPortValue()) // invalid listener port number
+	assert.Len(t, listener.FilterChains, 1)
+	chain := listener.FilterChains[0]
 	assert.Len(t, chain.Filters, 1)
 	assert.Equal(t, "envoy.filters.network.http_connection_manager", chain.Filters[0].Name)
 	message, err := chain.Filters[0].GetTypedConfig().UnmarshalNew()
@@ -503,12 +508,13 @@ func TestCiliumEnvoyConfigNoAddress(t *testing.T) {
 	resources, err := parser.ParseResources("namespace", "name", cec.Spec.Resources, false, false, false, true)
 	require.NoError(t, err)
 	assert.Len(t, resources.Listeners, 1)
-	assert.Equal(t, "namespace/name/envoy-prometheus-metrics-listener", resources.Listeners[0].Name)
-	assert.NotNil(t, resources.Listeners[0].Address)
-	assert.NotNil(t, resources.Listeners[0].Address.GetSocketAddress())
-	assert.NotEqual(t, 0, resources.Listeners[0].Address.GetSocketAddress().GetPortValue())
-	assert.Len(t, resources.Listeners[0].FilterChains, 1)
-	chain := resources.Listeners[0].FilterChains[0]
+	listener := resources.Listeners["namespace/name/envoy-prometheus-metrics-listener"]
+	assert.Equal(t, "namespace/name/envoy-prometheus-metrics-listener", listener.Name)
+	assert.NotNil(t, listener.Address)
+	assert.NotNil(t, listener.Address.GetSocketAddress())
+	assert.NotEqual(t, 0, listener.Address.GetSocketAddress().GetPortValue())
+	assert.Len(t, listener.FilterChains, 1)
+	chain := listener.FilterChains[0]
 	assert.Len(t, chain.Filters, 2)
 	assert.Equal(t, "cilium.network", chain.Filters[0].Name)
 	assert.Equal(t, "envoy.filters.network.http_connection_manager", chain.Filters[1].Name)
@@ -628,11 +634,12 @@ func TestCiliumEnvoyConfigMulti(t *testing.T) {
 	resources, err := parser.ParseResources("namespace", "name", cec.Spec.Resources, false, false, false, true)
 	require.NoError(t, err)
 	assert.Len(t, resources.Listeners, 1)
-	assert.Equal(t, "namespace/name/multi-resource-listener", resources.Listeners[0].Name)
-	assert.Nil(t, resources.Listeners[0].GetInternalListener())
-	assert.Equal(t, uint32(10000), resources.Listeners[0].Address.GetSocketAddress().GetPortValue())
-	assert.Len(t, resources.Listeners[0].FilterChains, 1)
-	chain := resources.Listeners[0].FilterChains[0]
+	listener := resources.Listeners["namespace/name/multi-resource-listener"]
+	assert.Equal(t, "namespace/name/multi-resource-listener", listener.Name)
+	assert.Nil(t, listener.GetInternalListener())
+	assert.Equal(t, uint32(10000), listener.Address.GetSocketAddress().GetPortValue())
+	assert.Len(t, listener.FilterChains, 1)
+	chain := listener.FilterChains[0]
 	assert.Len(t, chain.Filters, 1)
 	assert.Equal(t, "envoy.filters.network.http_connection_manager", chain.Filters[0].Name)
 	message, err := chain.Filters[0].GetTypedConfig().UnmarshalNew()
@@ -659,9 +666,10 @@ func TestCiliumEnvoyConfigMulti(t *testing.T) {
 	//
 	assert.Equal(t, "type.googleapis.com/envoy.config.route.v3.RouteConfiguration", cec.Spec.Resources[1].TypeUrl)
 	assert.Len(t, resources.Routes, 1)
-	assert.Equal(t, "namespace/name/local_route", resources.Routes[0].Name)
-	assert.Len(t, resources.Routes[0].VirtualHosts, 1)
-	vh := resources.Routes[0].VirtualHosts[0]
+	route := resources.Routes["namespace/name/local_route"]
+	assert.Equal(t, "namespace/name/local_route", route.Name)
+	assert.Len(t, route.VirtualHosts, 1)
+	vh := route.VirtualHosts[0]
 	assert.Equal(t, "namespace/name/local_service", vh.Name)
 	assert.Len(t, vh.Domains, 1)
 	assert.Equal(t, "*", vh.Domains[0])
@@ -676,15 +684,16 @@ func TestCiliumEnvoyConfigMulti(t *testing.T) {
 	//
 	assert.Equal(t, "type.googleapis.com/envoy.config.cluster.v3.Cluster", cec.Spec.Resources[2].TypeUrl)
 	assert.Len(t, resources.Clusters, 1)
-	assert.Equal(t, "namespace/name/some_service", resources.Clusters[0].Name)
-	assert.Equal(t, int64(0), resources.Clusters[0].ConnectTimeout.Seconds)
-	assert.Equal(t, int32(250000000), resources.Clusters[0].ConnectTimeout.Nanos)
-	assert.Equal(t, envoy_config_cluster.Cluster_ROUND_ROBIN, resources.Clusters[0].LbPolicy)
-	assert.Equal(t, envoy_config_cluster.Cluster_EDS, resources.Clusters[0].GetType())
+	cluster := resources.Clusters["namespace/name/some_service"]
+	assert.Equal(t, "namespace/name/some_service", cluster.Name)
+	assert.Equal(t, int64(0), cluster.ConnectTimeout.Seconds)
+	assert.Equal(t, int32(250000000), cluster.ConnectTimeout.Nanos)
+	assert.Equal(t, envoy_config_cluster.Cluster_ROUND_ROBIN, cluster.LbPolicy)
+	assert.Equal(t, envoy_config_cluster.Cluster_EDS, cluster.GetType())
 	//
 	// Check that missing CircuitBreakers is automatically filled in
 	//
-	cb := resources.Clusters[0].CircuitBreakers
+	cb := cluster.CircuitBreakers
 	assert.NotNil(t, cb)
 	assert.Len(t, cb.Thresholds, 1)
 	assert.Equal(t, uint32(128), cb.Thresholds[0].MaxRetries.Value)
@@ -693,13 +702,13 @@ func TestCiliumEnvoyConfigMulti(t *testing.T) {
 	//
 	// Check that missing EDS config source is automatically filled in
 	//
-	eds := resources.Clusters[0].GetEdsClusterConfig()
+	eds := cluster.GetEdsClusterConfig()
 	assert.NotNil(t, eds)
 	checkCiliumXDS(t, eds.GetEdsConfig())
 
-	assert.NotNil(t, resources.Clusters[0].TransportSocket)
-	assert.Equal(t, "envoy.transport_sockets.tls", resources.Clusters[0].TransportSocket.Name)
-	msg, err := resources.Clusters[0].TransportSocket.GetTypedConfig().UnmarshalNew()
+	assert.NotNil(t, cluster.TransportSocket)
+	assert.Equal(t, "envoy.transport_sockets.tls", cluster.TransportSocket.Name)
+	msg, err := cluster.TransportSocket.GetTypedConfig().UnmarshalNew()
 	require.NoError(t, err)
 	assert.NotNil(t, msg)
 	tls, ok := msg.(*envoy_config_tls.UpstreamTlsContext)
@@ -722,10 +731,11 @@ func TestCiliumEnvoyConfigMulti(t *testing.T) {
 	//
 	assert.Equal(t, "type.googleapis.com/envoy.config.endpoint.v3.ClusterLoadAssignment", cec.Spec.Resources[3].TypeUrl)
 	assert.Len(t, resources.Endpoints, 2)
-	assert.Equal(t, "namespace/name/some_service", resources.Endpoints[0].ClusterName)
-	assert.Len(t, resources.Endpoints[0].Endpoints, 1)
-	assert.Len(t, resources.Endpoints[0].Endpoints[0].LbEndpoints, 1)
-	addr := resources.Endpoints[0].Endpoints[0].LbEndpoints[0].GetEndpoint().Address
+	endpoint := resources.Endpoints["namespace/name/some_service"]
+	assert.Equal(t, "namespace/name/some_service", endpoint.ClusterName)
+	assert.Len(t, endpoint.Endpoints, 1)
+	assert.Len(t, endpoint.Endpoints[0].LbEndpoints, 1)
+	addr := endpoint.Endpoints[0].LbEndpoints[0].GetEndpoint().Address
 	assert.NotNil(t, addr)
 	assert.NotNil(t, addr.GetSocketAddress())
 	assert.Equal(t, "127.0.0.1", addr.GetSocketAddress().GetAddress())
@@ -736,10 +746,11 @@ func TestCiliumEnvoyConfigMulti(t *testing.T) {
 	//
 	assert.Equal(t, "type.googleapis.com/envoy.config.endpoint.v3.ClusterLoadAssignment", cec.Spec.Resources[4].TypeUrl)
 	assert.Len(t, resources.Endpoints, 2)
-	assert.Equal(t, "namespace/name/other_service", resources.Endpoints[1].ClusterName)
-	assert.Len(t, resources.Endpoints[1].Endpoints, 1)
-	assert.Len(t, resources.Endpoints[1].Endpoints[0].LbEndpoints, 1)
-	addr = resources.Endpoints[1].Endpoints[0].LbEndpoints[0].GetEndpoint().Address
+	other_endpoint := resources.Endpoints["namespace/name/other_service"]
+	assert.Equal(t, "namespace/name/other_service", other_endpoint.ClusterName)
+	assert.Len(t, other_endpoint.Endpoints, 1)
+	assert.Len(t, other_endpoint.Endpoints[0].LbEndpoints, 1)
+	addr = other_endpoint.Endpoints[0].LbEndpoints[0].GetEndpoint().Address
 	assert.NotNil(t, addr)
 	assert.NotNil(t, addr.GetSocketAddress())
 	assert.Equal(t, "::", addr.GetSocketAddress().GetAddress())
@@ -826,10 +837,11 @@ func TestCiliumEnvoyConfigInternalListener(t *testing.T) {
 	// Check internal endpoint resource
 	//
 	assert.Len(t, resources.Endpoints, 1)
-	assert.Equal(t, "namespace/name/internal_listener_cluster", resources.Endpoints[0].ClusterName)
-	assert.Len(t, resources.Endpoints[0].Endpoints, 1)
-	assert.Len(t, resources.Endpoints[0].Endpoints[0].LbEndpoints, 1)
-	addr := resources.Endpoints[0].Endpoints[0].LbEndpoints[0].GetEndpoint().Address
+	endpoint := resources.Endpoints["namespace/name/internal_listener_cluster"]
+	assert.Equal(t, "namespace/name/internal_listener_cluster", endpoint.ClusterName)
+	assert.Len(t, endpoint.Endpoints, 1)
+	assert.Len(t, endpoint.Endpoints[0].LbEndpoints, 1)
+	addr := endpoint.Endpoints[0].LbEndpoints[0].GetEndpoint().Address
 	assert.NotNil(t, addr)
 	assert.NotNil(t, addr.GetEnvoyInternalAddress())
 	assert.Equal(t, "namespace/name/internal-listener", addr.GetEnvoyInternalAddress().GetServerListenerName())
@@ -838,8 +850,9 @@ func TestCiliumEnvoyConfigInternalListener(t *testing.T) {
 	// Check internal listener
 	//
 	assert.Len(t, resources.Listeners, 1)
-	assert.Equal(t, "namespace/name/internal-listener", resources.Listeners[0].Name)
-	assert.NotNil(t, resources.Listeners[0].GetInternalListener())
+	listener := resources.Listeners["namespace/name/internal-listener"]
+	assert.Equal(t, "namespace/name/internal-listener", listener.Name)
+	assert.NotNil(t, listener.GetInternalListener())
 }
 
 var ciliumEnvoyConfigMissingInternalListener = `apiVersion: cilium.io/v2
@@ -911,7 +924,7 @@ spec:
 // The parser's enableBPFTProxy field is set directly on the struct so the test
 // does not have to mutate the global option.Config.EnableBPFTProxy.
 func TestCiliumEnvoyConfigReusePortWithInternalListener(t *testing.T) {
-	parseListeners := func(t *testing.T, enableBPFTProxy bool) []*envoy_config_listener.Listener {
+	parseListeners := func(t *testing.T, enableBPFTProxy bool) map[string]*envoy_config_listener.Listener {
 		t.Helper()
 
 		parser := CECResourceParser{
@@ -934,15 +947,11 @@ func TestCiliumEnvoyConfigReusePortWithInternalListener(t *testing.T) {
 		return resources.Listeners
 	}
 
-	findListener := func(t *testing.T, listeners []*envoy_config_listener.Listener, name string) *envoy_config_listener.Listener {
+	findListener := func(t *testing.T, listeners map[string]*envoy_config_listener.Listener, name string) *envoy_config_listener.Listener {
 		t.Helper()
-		for _, l := range listeners {
-			if l.Name == name {
-				return l
-			}
-		}
-		t.Fatalf("listener %q not found", name)
-		return nil
+		listener, ok := listeners[name]
+		require.Truef(t, ok, "listener %q not found", name)
+		return listener
 	}
 
 	t.Run("BPF TProxy enabled", func(t *testing.T) {
@@ -992,15 +1001,16 @@ func TestCiliumEnvoyConfigTCPProxy(t *testing.T) {
 	resources, err := parser.ParseResources("namespace", "name", cec.Spec.Resources, false, false, true, true)
 	require.NoError(t, err)
 	assert.Len(t, resources.Listeners, 1)
-	assert.NotNil(t, resources.Listeners[0].Address)
-	assert.NotNil(t, resources.Listeners[0].Address.GetSocketAddress())
-	assert.NotEqual(t, 0, resources.Listeners[0].Address.GetSocketAddress().GetPortValue())
+	listener := resources.Listeners["namespace/name/tcp_proxy_test-2"]
+	assert.NotNil(t, listener.Address)
+	assert.NotNil(t, listener.Address.GetSocketAddress())
+	assert.NotEqual(t, 0, listener.Address.GetSocketAddress().GetPortValue())
 	//
 	// Check injected listener filter config
 	//
-	assert.Len(t, resources.Listeners[0].ListenerFilters, 1)
-	assert.Equal(t, "cilium.bpf_metadata", resources.Listeners[0].ListenerFilters[0].Name)
-	lfMsg, err := resources.Listeners[0].ListenerFilters[0].GetTypedConfig().UnmarshalNew()
+	assert.Len(t, listener.ListenerFilters, 1)
+	assert.Equal(t, "cilium.bpf_metadata", listener.ListenerFilters[0].Name)
+	lfMsg, err := listener.ListenerFilters[0].GetTypedConfig().UnmarshalNew()
 	require.NoError(t, err)
 	assert.NotNil(t, lfMsg)
 	lf, ok := lfMsg.(*cilium.BpfMetadata)
@@ -1014,8 +1024,8 @@ func TestCiliumEnvoyConfigTCPProxy(t *testing.T) {
 	// TCP listener has no SO_LINGER config
 	assert.Nil(t, lf.OriginalSourceSoLingerTime)
 
-	assert.Len(t, resources.Listeners[0].FilterChains, 1)
-	chain := resources.Listeners[0].FilterChains[0]
+	assert.Len(t, listener.FilterChains, 1)
+	chain := listener.FilterChains[0]
 	assert.Len(t, chain.Filters, 2)
 	assert.Equal(t, "cilium.network", chain.Filters[0].Name)
 	assert.Equal(t, "envoy.filters.network.tcp_proxy", chain.Filters[1].Name)
@@ -1038,13 +1048,14 @@ func TestCiliumEnvoyConfigTCPProxy(t *testing.T) {
 	//
 	assert.Equal(t, "type.googleapis.com/envoy.config.cluster.v3.Cluster", cec.Spec.Resources[1].TypeUrl)
 	assert.Len(t, resources.Clusters, 1)
-	assert.Equal(t, "namespace/name/cluster_0", resources.Clusters[0].Name)
-	assert.Equal(t, int64(5), resources.Clusters[0].ConnectTimeout.Seconds)
-	assert.Equal(t, int32(0), resources.Clusters[0].ConnectTimeout.Nanos)
-	assert.Equal(t, "namespace/name/cluster_0", resources.Clusters[0].LoadAssignment.ClusterName)
-	assert.Len(t, resources.Clusters[0].LoadAssignment.Endpoints, 1)
-	assert.Len(t, resources.Clusters[0].LoadAssignment.Endpoints[0].LbEndpoints, 1)
-	addr := resources.Clusters[0].LoadAssignment.Endpoints[0].LbEndpoints[0].GetEndpoint().Address
+	cluster := resources.Clusters["namespace/name/cluster_0"]
+	assert.Equal(t, "namespace/name/cluster_0", cluster.Name)
+	assert.Equal(t, int64(5), cluster.ConnectTimeout.Seconds)
+	assert.Equal(t, int32(0), cluster.ConnectTimeout.Nanos)
+	assert.Equal(t, "namespace/name/cluster_0", cluster.LoadAssignment.ClusterName)
+	assert.Len(t, cluster.LoadAssignment.Endpoints, 1)
+	assert.Len(t, cluster.LoadAssignment.Endpoints[0].LbEndpoints, 1)
+	addr := cluster.LoadAssignment.Endpoints[0].LbEndpoints[0].GetEndpoint().Address
 	assert.NotNil(t, addr)
 	assert.NotNil(t, addr.GetSocketAddress())
 	assert.Equal(t, "127.0.0.1", addr.GetSocketAddress().GetAddress())
@@ -1135,15 +1146,16 @@ func TestCiliumEnvoyConfigTCPProxyTermination(t *testing.T) {
 	resources, err := parser.ParseResources("namespace", "name", cec.Spec.Resources, true, true, false, true)
 	require.NoError(t, err)
 	assert.Len(t, resources.Listeners, 1)
-	assert.NotNil(t, resources.Listeners[0].Address)
-	assert.NotNil(t, resources.Listeners[0].Address.GetSocketAddress())
-	assert.NotEqual(t, 0, resources.Listeners[0].Address.GetSocketAddress().GetPortValue())
+	listener := resources.Listeners["namespace/name/envoy-ingress-listener"]
+	assert.NotNil(t, listener.Address)
+	assert.NotNil(t, listener.Address.GetSocketAddress())
+	assert.NotEqual(t, 0, listener.Address.GetSocketAddress().GetPortValue())
 	//
 	// Check injected listener filter config
 	//
-	assert.Len(t, resources.Listeners[0].ListenerFilters, 1)
-	assert.Equal(t, "cilium.bpf_metadata", resources.Listeners[0].ListenerFilters[0].Name)
-	lfMsg, err := resources.Listeners[0].ListenerFilters[0].GetTypedConfig().UnmarshalNew()
+	assert.Len(t, listener.ListenerFilters, 1)
+	assert.Equal(t, "cilium.bpf_metadata", listener.ListenerFilters[0].Name)
+	lfMsg, err := listener.ListenerFilters[0].GetTypedConfig().UnmarshalNew()
 	require.NoError(t, err)
 	assert.NotNil(t, lfMsg)
 	lf, ok := lfMsg.(*cilium.BpfMetadata)
@@ -1158,8 +1170,9 @@ func TestCiliumEnvoyConfigTCPProxyTermination(t *testing.T) {
 	assert.NotNil(t, lf.OriginalSourceSoLingerTime)
 	assert.Zero(t, *lf.OriginalSourceSoLingerTime)
 
-	assert.Len(t, resources.Listeners[0].FilterChains, 1)
-	chain := resources.Listeners[0].FilterChains[0]
+	listener = resources.Listeners["namespace/name/envoy-ingress-listener"]
+	assert.Len(t, listener.FilterChains, 1)
+	chain := listener.FilterChains[0]
 	assert.Len(t, chain.Filters, 2)
 	assert.Equal(t, "cilium.network", chain.Filters[0].Name)
 	assert.Equal(t, "envoy.filters.network.http_connection_manager", chain.Filters[1].Name)
@@ -1183,17 +1196,18 @@ func TestCiliumEnvoyConfigTCPProxyTermination(t *testing.T) {
 	//
 	assert.Equal(t, "type.googleapis.com/envoy.config.cluster.v3.Cluster", cec.Spec.Resources[1].TypeUrl)
 	assert.Len(t, resources.Clusters, 1)
-	assert.Equal(t, "default/service_google", resources.Clusters[0].Name)
-	assert.Equal(t, int64(5), resources.Clusters[0].ConnectTimeout.Seconds)
-	assert.Equal(t, int32(0), resources.Clusters[0].ConnectTimeout.Nanos)
-	assert.Equal(t, envoy_config_cluster.Cluster_LOGICAL_DNS, resources.Clusters[0].GetType())
-	assert.Equal(t, envoy_config_cluster.Cluster_V4_ONLY, resources.Clusters[0].GetDnsLookupFamily())
-	assert.Equal(t, envoy_config_cluster.Cluster_ROUND_ROBIN, resources.Clusters[0].LbPolicy)
+	cluster := resources.Clusters["default/service_google"]
+	assert.Equal(t, "default/service_google", cluster.Name)
+	assert.Equal(t, int64(5), cluster.ConnectTimeout.Seconds)
+	assert.Equal(t, int32(0), cluster.ConnectTimeout.Nanos)
+	assert.Equal(t, envoy_config_cluster.Cluster_LOGICAL_DNS, cluster.GetType())
+	assert.Equal(t, envoy_config_cluster.Cluster_V4_ONLY, cluster.GetDnsLookupFamily())
+	assert.Equal(t, envoy_config_cluster.Cluster_ROUND_ROBIN, cluster.LbPolicy)
 
-	assert.Equal(t, "default/service_google", resources.Clusters[0].LoadAssignment.ClusterName)
-	assert.Len(t, resources.Clusters[0].LoadAssignment.Endpoints, 1)
-	assert.Len(t, resources.Clusters[0].LoadAssignment.Endpoints[0].LbEndpoints, 1)
-	addr := resources.Clusters[0].LoadAssignment.Endpoints[0].LbEndpoints[0].GetEndpoint().Address
+	assert.Equal(t, "default/service_google", cluster.LoadAssignment.ClusterName)
+	assert.Len(t, cluster.LoadAssignment.Endpoints, 1)
+	assert.Len(t, cluster.LoadAssignment.Endpoints[0].LbEndpoints, 1)
+	addr := cluster.LoadAssignment.Endpoints[0].LbEndpoints[0].GetEndpoint().Address
 	assert.NotNil(t, addr)
 	assert.NotNil(t, addr.GetSocketAddress())
 	assert.Equal(t, "www.google.com", addr.GetSocketAddress().GetAddress())
@@ -1201,10 +1215,10 @@ func TestCiliumEnvoyConfigTCPProxyTermination(t *testing.T) {
 	//
 	// Check upstream filters (injected for L7 LB)
 	//
-	assert.NotNil(t, resources.Clusters[0].TypedExtensionProtocolOptions)
-	assert.NotNil(t, resources.Clusters[0].TypedExtensionProtocolOptions[httpProtocolOptionsType])
+	assert.NotNil(t, cluster.TypedExtensionProtocolOptions)
+	assert.NotNil(t, cluster.TypedExtensionProtocolOptions[httpProtocolOptionsType])
 	opts := &envoy_upstreams_http_v3.HttpProtocolOptions{}
-	assert.NoError(t, resources.Clusters[0].TypedExtensionProtocolOptions[httpProtocolOptionsType].UnmarshalTo(opts))
+	assert.NoError(t, cluster.TypedExtensionProtocolOptions[httpProtocolOptionsType].UnmarshalTo(opts))
 	assert.NotNil(t, opts.HttpFilters)
 	assert.Equal(t, "cilium.l7policy", opts.HttpFilters[0].Name)
 	assert.Equal(t, ciliumL7FilterTypeURL, opts.HttpFilters[0].GetTypedConfig().TypeUrl)
@@ -1269,7 +1283,8 @@ func TestCiliumEnvoyConfigtHTTPHealthCheckFilter(t *testing.T) {
 	resources, err := parser.ParseResources(cec.Namespace, cec.Name, cec.Spec.Resources, false, false, false, false)
 	require.NoError(t, err)
 	assert.Len(t, resources.Listeners, 1)
-	chain := resources.Listeners[0].FilterChains[0]
+	listener := resources.Listeners["test-namespace/test-name/listener"]
+	chain := listener.FilterChains[0]
 	assert.Len(t, chain.Filters, 1)
 	assert.Equal(t, "envoy.filters.network.http_connection_manager", chain.Filters[0].Name)
 
@@ -1287,22 +1302,23 @@ func TestCiliumEnvoyConfigtHTTPHealthCheckFilter(t *testing.T) {
 }
 
 func TestListenersAddedOrDeleted(t *testing.T) {
-	var old envoy.Resources
-	var new envoy.Resources
+	old := xds.NewResources()
+	new := xds.NewResources()
 
 	// Both empty
 	res := old.ListenersAddedOrDeleted(&new)
 	assert.False(t, res)
 
 	// new adds a listener
-	new.Listeners = append(old.Listeners, &envoy_config_listener.Listener{Name: "foo"})
+	maps.Copy(new.Listeners, old.Listeners)
+	new.Listeners["foo"] = &envoy_config_listener.Listener{Name: "foo"}
 	res = old.ListenersAddedOrDeleted(&new)
 	assert.True(t, res)
 	res = new.ListenersAddedOrDeleted(&old)
 	assert.True(t, res)
 
 	// Now both have 'foo'
-	old.Listeners = append(old.Listeners, &envoy_config_listener.Listener{Name: "foo"})
+	old.Listeners["foo"] = &envoy_config_listener.Listener{Name: "foo"}
 	res = old.ListenersAddedOrDeleted(&new)
 	assert.False(t, res)
 	res = new.ListenersAddedOrDeleted(&old)
@@ -1316,21 +1332,22 @@ func TestListenersAddedOrDeleted(t *testing.T) {
 	assert.True(t, res)
 
 	// New has a different listener
-	new.Listeners = append(new.Listeners, &envoy_config_listener.Listener{Name: "bar"})
+	new.Listeners = make(map[string]*envoy_config_listener.Listener)
+	new.Listeners["bar"] = &envoy_config_listener.Listener{Name: "bar"}
 	res = old.ListenersAddedOrDeleted(&new)
 	assert.True(t, res)
 	res = new.ListenersAddedOrDeleted(&old)
 	assert.True(t, res)
 
 	// New adds the listener in old, but still has the other listener
-	new.Listeners = append(new.Listeners, &envoy_config_listener.Listener{Name: "foo"})
+	new.Listeners["foo"] = &envoy_config_listener.Listener{Name: "foo"}
 	res = old.ListenersAddedOrDeleted(&new)
 	assert.True(t, res)
 	res = new.ListenersAddedOrDeleted(&old)
 	assert.True(t, res)
 
 	// Same listeners but in different order
-	old.Listeners = append(old.Listeners, &envoy_config_listener.Listener{Name: "bar"})
+	old.Listeners["bar"] = &envoy_config_listener.Listener{Name: "bar"}
 	res = old.ListenersAddedOrDeleted(&new)
 	assert.False(t, res)
 	res = new.ListenersAddedOrDeleted(&old)
@@ -1400,10 +1417,11 @@ func TestCiliumEnvoyConfigCombinedValidationContext(t *testing.T) {
 	require.NoError(t, err)
 
 	require.Len(t, resources.Listeners, 1)
-	assert.Equal(t, "namespace/name/combined-validationcontext", resources.Listeners[0].Name)
-	assert.Equal(t, uint32(10000), resources.Listeners[0].Address.GetSocketAddress().GetPortValue())
-	assert.Len(t, resources.Listeners[0].FilterChains, 1)
-	chain := resources.Listeners[0].FilterChains[0]
+	listener := resources.Listeners["namespace/name/combined-validationcontext"]
+	assert.Equal(t, "namespace/name/combined-validationcontext", listener.Name)
+	assert.Equal(t, uint32(10000), listener.Address.GetSocketAddress().GetPortValue())
+	assert.Len(t, listener.FilterChains, 1)
+	chain := listener.FilterChains[0]
 
 	assert.NotNil(t, chain.TransportSocket)
 	assert.Equal(t, "envoy.transport_sockets.tls", chain.TransportSocket.Name)
@@ -1486,10 +1504,11 @@ func TestCiliumEnvoyConfigTlsSessionTicketKeys(t *testing.T) {
 	require.NoError(t, err)
 
 	require.Len(t, resources.Listeners, 1)
-	assert.Equal(t, "namespace/name/sessionticket-keys", resources.Listeners[0].Name)
-	assert.Equal(t, uint32(10000), resources.Listeners[0].Address.GetSocketAddress().GetPortValue())
-	assert.Len(t, resources.Listeners[0].FilterChains, 1)
-	chain := resources.Listeners[0].FilterChains[0]
+	listener := resources.Listeners["namespace/name/sessionticket-keys"]
+	assert.Equal(t, "namespace/name/sessionticket-keys", listener.Name)
+	assert.Equal(t, uint32(10000), listener.Address.GetSocketAddress().GetPortValue())
+	assert.Len(t, listener.FilterChains, 1)
+	chain := listener.FilterChains[0]
 
 	assert.NotNil(t, chain.TransportSocket)
 	assert.Equal(t, "envoy.transport_sockets.tls", chain.TransportSocket.Name)
@@ -1561,9 +1580,11 @@ func TestCiliumEnvoyConfigInjectCiliumFilters(t *testing.T) {
 	require.NoError(t, err)
 
 	require.Len(t, resources.Listeners, 1)
-	assert.Equal(t, "namespace/name/without-cilium-filters", resources.Listeners[0].Name)
-	assert.Len(t, resources.Listeners[0].FilterChains, 1)
-	chain := resources.Listeners[0].FilterChains[0]
+	listener := resources.Listeners["namespace/name/without-cilium-filters"]
+	require.NotNil(t, listener)
+	assert.Equal(t, "namespace/name/without-cilium-filters", listener.Name)
+	assert.Len(t, listener.FilterChains, 1)
+	chain := listener.FilterChains[0]
 
 	//
 	// Check that missing Cilium Envoy filters are not automatically filled in
@@ -1637,10 +1658,12 @@ func TestParseEnvoySpec(t *testing.T) {
 	resources, err := parser.ParseResources("", "name", cec.Spec.Resources, len(cec.Spec.Services) > 0, InjectCiliumEnvoyFilters(&cec.ObjectMeta, &cec.Spec), UseOriginalSourceAddress(&cec.ObjectMeta), true)
 	assert.NoError(t, err)
 	assert.Len(t, resources.Listeners, 1)
-	assert.Equal(t, uint32(10000), resources.Listeners[0].Address.GetSocketAddress().GetPortValue())
-	assert.Len(t, resources.Listeners[0].FilterChains, 1)
-	assert.Equal(t, "/name/envoy-prometheus-metrics-listener", resources.Listeners[0].Name)
-	chain := resources.Listeners[0].FilterChains[0]
+	listener := resources.Listeners["/name/envoy-prometheus-metrics-listener"]
+	require.NotNil(t, listener)
+	assert.Equal(t, uint32(10000), listener.Address.GetSocketAddress().GetPortValue())
+	assert.Len(t, listener.FilterChains, 1)
+	assert.Equal(t, "/name/envoy-prometheus-metrics-listener", listener.Name)
+	chain := listener.FilterChains[0]
 	assert.Len(t, chain.Filters, 1)
 	assert.Equal(t, "envoy.filters.network.http_connection_manager", chain.Filters[0].Name)
 	message, err := chain.Filters[0].GetTypedConfig().UnmarshalNew()
@@ -1715,10 +1738,12 @@ func TestParseEnvoySpecWithService(t *testing.T) {
 	resources, err := parser.ParseResources("", "name", cec.Spec.Resources, len(cec.Spec.Services) > 0, InjectCiliumEnvoyFilters(&cec.ObjectMeta, &cec.Spec), UseOriginalSourceAddress(&cec.ObjectMeta), true)
 	assert.NoError(t, err)
 	assert.Len(t, resources.Listeners, 1)
-	assert.Equal(t, uint32(1025), resources.Listeners[0].Address.GetSocketAddress().GetPortValue())
-	assert.Len(t, resources.Listeners[0].FilterChains, 1)
-	assert.Equal(t, "/name/l7-lb", resources.Listeners[0].Name)
-	chain := resources.Listeners[0].FilterChains[0]
+	listener := resources.Listeners["/name/l7-lb"]
+	require.NotNil(t, listener)
+	assert.Equal(t, uint32(1025), listener.Address.GetSocketAddress().GetPortValue())
+	assert.Len(t, listener.FilterChains, 1)
+	assert.Equal(t, "/name/l7-lb", listener.Name)
+	chain := listener.FilterChains[0]
 	assert.Len(t, chain.Filters, 2)
 	assert.Equal(t, "cilium.network", chain.Filters[0].Name)
 	assert.Equal(t, "envoy.filters.network.http_connection_manager", chain.Filters[1].Name)
