+++ pkg/ciliumenvoyconfig/controller.go
@@ -357,7 +357,7 @@ func (bs *backendProcessor) process(wtxn statedb.WriteTxn, closedWatches []<-cha
 		}
 
 		prevEndpoints := res.Resources.Endpoints
-		var newEndpoints []*envoy_config_endpoint.ClusterLoadAssignment
+		var newEndpoints map[string]*envoy_config_endpoint.ClusterLoadAssignment
 
 		// Look up the referenced service for the port name to port number mappings.
 		svc, _, watchSvc, found := bs.writer.Services().GetWatch(wtxn, loadbalancer.ServiceByName(res.ClusterServiceName()))
@@ -366,11 +366,14 @@ func (bs *backendProcessor) process(wtxn statedb.WriteTxn, closedWatches []<-cha
 			// Look up associated backends and update the load assignments.
 			bes, watchBes := bs.writer.BackendsForService(wtxn, svc.Name)
 			ws.Add(watchBes)
-			newEndpoints = computeLoadAssignments(
+			newEndpoints = make(map[string]*envoy_config_endpoint.ClusterLoadAssignment)
+			for _, assignment := range computeLoadAssignments(
 				svc.Name,
 				res.ClusterReferences,
 				svc.PortNames,
-				bs.writer.SelectBackends(wtxn, bes, svc, nil))
+				bs.writer.SelectBackends(wtxn, bes, svc, nil)) {
+				newEndpoints[assignment.ClusterName] = assignment
+			}
 		} else {
 			// No service found (yet) and thus there are no endpoints.
 			newEndpoints = nil
@@ -379,7 +382,7 @@ func (bs *backendProcessor) process(wtxn statedb.WriteTxn, closedWatches []<-cha
 		claEqual := func(a, b *envoy_config_endpoint.ClusterLoadAssignment) bool {
 			return proto.Equal(a, b)
 		}
-		endpointsEqual := slices.EqualFunc(prevEndpoints, newEndpoints, claEqual)
+		endpointsEqual := maps.EqualFunc(prevEndpoints, newEndpoints, claEqual)
 
 		if !endpointsEqual || (res.Status.Kind != reconciler.StatusKindDone && res.Status.Kind != reconciler.StatusKindPending) {
 			res = res.Clone()
