+++ pkg/ciliumenvoyconfig/script_test.go
@@ -38,8 +38,8 @@ import (
 	cmtypes "github.com/cilium/cilium/pkg/clustermesh/types"
 	"github.com/cilium/cilium/pkg/datapath/tables"
 	"github.com/cilium/cilium/pkg/endpoint/regeneration"
-	"github.com/cilium/cilium/pkg/envoy"
 	envoyCfg "github.com/cilium/cilium/pkg/envoy/config"
+	"github.com/cilium/cilium/pkg/envoy/xds"
 	"github.com/cilium/cilium/pkg/hive"
 	k8sClient "github.com/cilium/cilium/pkg/k8s/client/testutils"
 	"github.com/cilium/cilium/pkg/k8s/synced"
@@ -290,7 +290,7 @@ func (rs resourceStore) equal(other resourceStore) bool {
 	return maps.EqualFunc(rs, other, proto.Equal)
 }
 
-func (rs resourceStore) update(res *envoy.Resources) {
+func (rs resourceStore) update(res *xds.Resources) {
 	for _, l := range res.Listeners {
 		rs[resourceKey{kind: listenerKey, name: l.Name}.String()] = l
 	}
@@ -308,7 +308,7 @@ func (rs resourceStore) update(res *envoy.Resources) {
 	}
 }
 
-func (rs resourceStore) delete(res *envoy.Resources) {
+func (rs resourceStore) delete(res *xds.Resources) {
 	for _, l := range res.Listeners {
 		delete(rs, resourceKey{kind: listenerKey, name: l.Name}.String())
 	}
@@ -455,7 +455,7 @@ func indentLines(s string) string {
 }
 
 // DeleteResources implements envoySyncer.
-func (f *fakeEnvoySyncerAndPolicyTrigger) DeleteEnvoyResources(ctx context.Context, res envoy.Resources) error {
+func (f *fakeEnvoySyncerAndPolicyTrigger) DeleteEnvoyResources(ctx context.Context, res xds.Resources) error {
 	f.Lock()
 	defer f.Unlock()
 	f.store.delete(&res)
@@ -469,7 +469,7 @@ func (f *fakeEnvoySyncerAndPolicyTrigger) DeleteEnvoyResources(ctx context.Conte
 }
 
 // UpdateResources implements envoySyncer.
-func (f *fakeEnvoySyncerAndPolicyTrigger) UpdateEnvoyResources(ctx context.Context, old envoy.Resources, new envoy.Resources) error {
+func (f *fakeEnvoySyncerAndPolicyTrigger) UpdateEnvoyResources(ctx context.Context, old xds.Resources, new xds.Resources) error {
 	f.Lock()
 	defer f.Unlock()
 	f.store.delete(&old)
