+++ pkg/endpoint/proxy.go
@@ -19,8 +19,8 @@ type EndpointProxy interface {
 	CreateOrUpdateRedirect(ctx context.Context, l4 policy.ProxyPolicy, id string, epID uint16, wg *completion.WaitGroup) (proxyPort uint16, err error, revertFunc revert.RevertFunc)
 	RemoveRedirect(id string)
 	UpdateSDP(rules map[identity.NumericIdentity]policy.SelectorPolicy)
-	UpdateNetworkPolicy(ep endpoint.EndpointUpdater, policy *policy.EndpointPolicy, wg *completion.WaitGroup) (error, revert.RevertFunc, revert.FinalizeFunc)
-	RemoveNetworkPolicy(ep endpoint.EndpointInfoSource)
+	UpdateNetworkPolicy(ctx context.Context, ep endpoint.EndpointUpdater, policy *policy.EndpointPolicy, wg *completion.WaitGroup) (error, revert.RevertFunc, revert.FinalizeFunc)
+	RemoveNetworkPolicy(ctx context.Context, ep endpoint.EndpointInfoSource)
 	GetListenerProxyPort(listener string) uint16
 	IsSDPEnabled() bool
 }
@@ -29,7 +29,7 @@ func (e *Endpoint) removeNetworkPolicy() {
 	if e.IsProxyDisabled() {
 		return
 	}
-	e.proxy.RemoveNetworkPolicy(e)
+	e.proxy.RemoveNetworkPolicy(context.Background(), e)
 }
 
 func (e *Endpoint) IsProxyDisabled() bool {
@@ -49,12 +49,13 @@ func (f *FakeEndpointProxy) RemoveRedirect(id string) {
 }
 
 // UpdateNetworkPolicy does nothing.
-func (f *FakeEndpointProxy) UpdateNetworkPolicy(ep endpoint.EndpointUpdater, policy *policy.EndpointPolicy, wg *completion.WaitGroup) (error, revert.RevertFunc, revert.FinalizeFunc) {
+func (f *FakeEndpointProxy) UpdateNetworkPolicy(ctx context.Context, ep endpoint.EndpointUpdater, policy *policy.EndpointPolicy, wg *completion.WaitGroup) (error, revert.RevertFunc, revert.FinalizeFunc) {
 	return nil, nil, nil
 }
 
 // RemoveNetworkPolicy does nothing.
-func (f *FakeEndpointProxy) RemoveNetworkPolicy(ep endpoint.EndpointInfoSource) {}
+func (f *FakeEndpointProxy) RemoveNetworkPolicy(ctx context.Context, ep endpoint.EndpointInfoSource) {
+}
 
 func (f *FakeEndpointProxy) UpdateSDP(rules map[identity.NumericIdentity]policy.SelectorPolicy) {
 }
