+++ pkg/envoy/accesslog_server.go
@@ -17,6 +17,8 @@ import (
 	"golang.org/x/sys/unix"
 	"google.golang.org/protobuf/proto"
 
+	util "github.com/cilium/cilium/pkg/envoy/util"
+
 	"github.com/cilium/cilium/pkg/flowdebug"
 	"github.com/cilium/cilium/pkg/identity"
 	"github.com/cilium/cilium/pkg/logging/logfields"
@@ -38,7 +40,7 @@ func newAccessLogServer(logger *slog.Logger, accessLogger accesslog.ProxyAccessL
 	return &AccessLogServer{
 		logger:             logger,
 		accessLogger:       accessLogger,
-		socketPath:         getAccessLogSocketPath(envoySocketDir),
+		socketPath:         util.GetAccessLogSocketPath(),
 		proxyGID:           proxyGID,
 		localEndpointStore: localEndpointStore,
 		bufferSize:         bufferSize,
