+++ pkg/envoy/resources.go
@@ -21,6 +21,7 @@ import (
 	"github.com/cilium/cilium/pkg/logging/logfields"
 )
 
+// todo (nezdolik) migrate to go control plane constants when available
 const (
 	// ListenerTypeURL is the type URL of Listener resources.
 	ListenerTypeURL = "type.googleapis.com/envoy.config.listener.v3.Listener"
