+++ pkg/envoy/xds_server_test.go
@@ -4,9 +4,11 @@
 package envoy
 
 import (
+	"context"
 	"iter"
 	"strconv"
 	"testing"
+	"time"
 
 	"github.com/cilium/hive/hivetest"
 	cilium "github.com/cilium/proxy/go/cilium/api"
@@ -18,10 +20,12 @@ import (
 	"github.com/stretchr/testify/require"
 	k8sTypes "k8s.io/apimachinery/pkg/types"
 
+	"github.com/cilium/cilium/pkg/completion"
 	"github.com/cilium/cilium/pkg/crypto/certificatemanager"
 	"github.com/cilium/cilium/pkg/endpoint/regeneration"
 	envoypolicy "github.com/cilium/cilium/pkg/envoy/policy"
 	"github.com/cilium/cilium/pkg/envoy/test"
+	"github.com/cilium/cilium/pkg/envoy/xds"
 	"github.com/cilium/cilium/pkg/identity"
 	"github.com/cilium/cilium/pkg/identity/identitymanager"
 	k8sConst "github.com/cilium/cilium/pkg/k8s/apis/cilium.io"
@@ -90,6 +94,12 @@ func newProxyUpdaterMock(t testing.TB, namedPort uint16) *test.ProxyUpdaterMock 
 type dummyPolicyStats struct {
 	waitingForPolicyRepository spanstat.SpanStat
 }
+
+type noopXDSMetrics struct{}
+
+func (noopXDSMetrics) IncreaseNACK(string)   {}
+func (noopXDSMetrics) IncreaseACK(string)    {}
+func (noopXDSMetrics) IncreaseCancel(string) {}
 
 func (s *dummyPolicyStats) WaitingForPolicyRepository() *spanstat.SpanStat {
 	return &s.waitingForPolicyRepository
@@ -740,7 +750,7 @@ func Test_getWildcardNetworkPolicyRules(t *testing.T) {
 			wildcardCachedSelector:    nil,
 		}
 
-		obtained, isPass, wildcardSelectorPrecedence := xds.getWildcardPortNetworkPolicyRules(ep, version, policyTypes.HighestPriority, policyTypes.LowestPriority, perSelectorPoliciesWithWildcard, false, false, "")
+		obtained, isPass, wildcardSelectorPrecedence := GetWildcardPortNetworkPolicyRules(ep, version, policyTypes.HighestPriority, policyTypes.LowestPriority, perSelectorPoliciesWithWildcard, false, false, "", xds.logger, xds.l7RulesTranslator)
 		require.Equal(t, []*cilium.PortNetworkPolicyRule{{
 			Precedence: uint32(policyTypes.MaxAllowPrecedence),
 		}}, obtained)
@@ -757,7 +767,7 @@ func Test_getWildcardNetworkPolicyRules(t *testing.T) {
 			cachedRequiresV2Selector1: nil,
 		}
 
-		obtained, isPass, wildcardSelectorPrecedence := xds.getWildcardPortNetworkPolicyRules(ep, version, policyTypes.HighestPriority, policyTypes.LowestPriority, perSelectorPolicies, false, false, "")
+		obtained, isPass, wildcardSelectorPrecedence := GetWildcardPortNetworkPolicyRules(ep, version, policyTypes.HighestPriority, policyTypes.LowestPriority, perSelectorPolicies, false, false, "", xds.logger, xds.l7RulesTranslator)
 		require.Equal(t, []*cilium.PortNetworkPolicyRule{{
 			Precedence:     uint32(policyTypes.MaxDenyPrecedence),
 			Verdict:        DenyVerdict,
@@ -777,9 +787,9 @@ func Test_getWildcardNetworkPolicyRules(t *testing.T) {
 			Verdict:  types.Pass,
 		}
 
-		obtained, isPass, wildcardSelectorPrecedence := xds.getWildcardPortNetworkPolicyRules(ep, version, policyTypes.HighestPriority, policyTypes.LowestPriority, policy.L7DataMap{
+		obtained, isPass, wildcardSelectorPrecedence := GetWildcardPortNetworkPolicyRules(ep, version, policyTypes.HighestPriority, policyTypes.LowestPriority, policy.L7DataMap{
 			wildcardCachedSelector: passPolicy,
-		}, false, false, "")
+		}, false, false, "", xds.logger, xds.l7RulesTranslator)
 
 		require.Equal(t, []*cilium.PortNetworkPolicyRule{{
 			Precedence: uint32(passPriority.ToPassPrecedence()),
@@ -797,7 +807,7 @@ func Test_getWildcardNetworkPolicyRules(t *testing.T) {
 		allowPriority := passPriority
 		denyPriority := passPriority
 
-		obtained, isPass, wildcardSelectorPrecedence := xds.getWildcardPortNetworkPolicyRules(ep, version, policyTypes.HighestPriority, policyTypes.LowestPriority, policy.L7DataMap{
+		obtained, isPass, wildcardSelectorPrecedence := GetWildcardPortNetworkPolicyRules(ep, version, policyTypes.HighestPriority, policyTypes.LowestPriority, policy.L7DataMap{
 			wildcardCachedSelector: {
 				Priority: passPriority,
 				Verdict:  types.Pass,
@@ -809,7 +819,7 @@ func Test_getWildcardNetworkPolicyRules(t *testing.T) {
 				Priority: denyPriority,
 				Verdict:  types.Deny,
 			},
-		}, false, false, "")
+		}, false, false, "", xds.logger, xds.l7RulesTranslator)
 
 		require.True(t, isPass)
 		require.Equal(t, passPriority.ToPassPrecedence(), wildcardSelectorPrecedence)
@@ -834,7 +844,7 @@ func Test_getWildcardNetworkPolicyRules(t *testing.T) {
 	t.Run("grouped_non_wildcard_pass_with_empty_selection_is_skipped", func(t *testing.T) {
 		noneCachedSelector, _ := testSelectorCache.AddIdentitySelectorForTest(dummySelectorCacheUser, api.EndpointSelectorNone)
 
-		obtained, isPass, wildcardSelectorPrecedence := xds.getWildcardPortNetworkPolicyRules(ep, version, policyTypes.HighestPriority, policyTypes.LowestPriority, policy.L7DataMap{
+		obtained, isPass, wildcardSelectorPrecedence := GetWildcardPortNetworkPolicyRules(ep, version, policyTypes.HighestPriority, policyTypes.LowestPriority, policy.L7DataMap{
 			cachedSelector1: {
 				Priority: policyTypes.Priority(3),
 			},
@@ -842,7 +852,7 @@ func Test_getWildcardNetworkPolicyRules(t *testing.T) {
 				Priority: policyTypes.Priority(2),
 				Verdict:  types.Pass,
 			},
-		}, false, false, "")
+		}, false, false, "", xds.logger, xds.l7RulesTranslator)
 
 		require.Equal(t, []*cilium.PortNetworkPolicyRule{{
 			Precedence:     uint32(policyTypes.Priority(3).ToAllowPrecedence()),
@@ -858,45 +868,45 @@ func TestGetPortNetworkPolicyRule(t *testing.T) {
 
 	version := testSelectorCache.GetSelectorSnapshot()
 
-	obtained, canShortCircuit := xds.getPortNetworkPolicyRule(ep, version, cachedSelector1, L7Rules12, policyTypes.LowestPriority, policyTypes.LowestPriority, false, false, "")
+	obtained, canShortCircuit := GetPortNetworkPolicyRule(ep, version, cachedSelector1, L7Rules12, policyTypes.LowestPriority, policyTypes.LowestPriority, false, false, "", xds.l7RulesTranslator, xds.logger)
 	require.Equal(t, ExpectedPortNetworkPolicyRule12, obtained)
 	require.True(t, canShortCircuit)
 
-	obtained, canShortCircuit = xds.getPortNetworkPolicyRule(ep, version, cachedSelector1, L7Rules12Deny, policyTypes.LowestPriority, policyTypes.LowestPriority, false, false, "")
+	obtained, canShortCircuit = GetPortNetworkPolicyRule(ep, version, cachedSelector1, L7Rules12Deny, policyTypes.LowestPriority, policyTypes.LowestPriority, false, false, "", xds.l7RulesTranslator, xds.logger)
 	require.Equal(t, ExpectedPortNetworkPolicyRule12Deny, obtained)
 	require.False(t, canShortCircuit)
 
-	obtained, canShortCircuit = xds.getPortNetworkPolicyRule(ep, version, cachedSelector1, L7Rules12HeaderMatch, policyTypes.LowestPriority, policyTypes.LowestPriority, false, false, "")
+	obtained, canShortCircuit = GetPortNetworkPolicyRule(ep, version, cachedSelector1, L7Rules12HeaderMatch, policyTypes.LowestPriority, policyTypes.LowestPriority, false, false, "", xds.l7RulesTranslator, xds.logger)
 	require.Equal(t, ExpectedPortNetworkPolicyRule122HeaderMatch, obtained)
 	require.False(t, canShortCircuit)
 
-	obtained, canShortCircuit = xds.getPortNetworkPolicyRule(ep, version, cachedSelector2, L7Rules1, policyTypes.LowestPriority, policyTypes.LowestPriority, false, false, "")
+	obtained, canShortCircuit = GetPortNetworkPolicyRule(ep, version, cachedSelector2, L7Rules1, policyTypes.LowestPriority, policyTypes.LowestPriority, false, false, "", xds.l7RulesTranslator, xds.logger)
 	require.Equal(t, ExpectedPortNetworkPolicyRule1, obtained)
 	require.True(t, canShortCircuit)
 
 	// With precedence
 
-	obtained, canShortCircuit = xds.getPortNetworkPolicyRule(ep, version, cachedSelector1, L7Rules12, policyTypes.HighestPriority, policyTypes.LowestPriority, false, false, "")
+	obtained, canShortCircuit = GetPortNetworkPolicyRule(ep, version, cachedSelector1, L7Rules12, policyTypes.HighestPriority, policyTypes.LowestPriority, false, false, "", xds.l7RulesTranslator, xds.logger)
 	require.Equal(t, ExpectedPortNetworkPolicyRule12Precedence, obtained)
 	require.True(t, canShortCircuit)
 
-	obtained, canShortCircuit = xds.getPortNetworkPolicyRule(ep, version, cachedSelector1, L7Rules12Deny, policyTypes.HighestPriority, policyTypes.LowestPriority, false, false, "")
+	obtained, canShortCircuit = GetPortNetworkPolicyRule(ep, version, cachedSelector1, L7Rules12Deny, policyTypes.HighestPriority, policyTypes.LowestPriority, false, false, "", xds.l7RulesTranslator, xds.logger)
 	require.Equal(t, ExpectedPortNetworkPolicyRule12DenyPrecedence, obtained)
 	require.False(t, canShortCircuit)
 
-	obtained, canShortCircuit = xds.getPortNetworkPolicyRule(ep, version, cachedSelector1, L7Rules12HeaderMatch, policyTypes.HighestPriority, policyTypes.LowestPriority, false, false, "")
+	obtained, canShortCircuit = GetPortNetworkPolicyRule(ep, version, cachedSelector1, L7Rules12HeaderMatch, policyTypes.HighestPriority, policyTypes.LowestPriority, false, false, "", xds.l7RulesTranslator, xds.logger)
 	require.Equal(t, ExpectedPortNetworkPolicyRule122HeaderMatchPrecedence, obtained)
 	require.False(t, canShortCircuit)
 
-	obtained, canShortCircuit = xds.getPortNetworkPolicyRule(ep, version, cachedSelector2, L7Rules1, policyTypes.HighestPriority, policyTypes.LowestPriority, false, false, "")
+	obtained, canShortCircuit = GetPortNetworkPolicyRule(ep, version, cachedSelector2, L7Rules1, policyTypes.HighestPriority, policyTypes.LowestPriority, false, false, "", xds.l7RulesTranslator, xds.logger)
 	require.Equal(t, ExpectedPortNetworkPolicyRule1Precedence, obtained)
 	require.True(t, canShortCircuit)
 
 	// with pass verdict
 
-	obtained, canShortCircuit = xds.getPortNetworkPolicyRule(ep, version, cachedSelector1,
+	obtained, canShortCircuit = GetPortNetworkPolicyRule(ep, version, cachedSelector1,
 		&policy.PerSelectorPolicy{Verdict: types.Pass, Priority: 0xffff},
-		0xffff, 0x1ffff, false, false, "")
+		0xffff, 0x1ffff, false, false, "", xds.l7RulesTranslator, xds.logger)
 	require.Equal(t, &cilium.PortNetworkPolicyRule{
 		Precedence:     0xff000000,
 		Verdict:        &cilium.PortNetworkPolicyRule_PassPrecedence{PassPrecedence: 0xfe000000},
@@ -909,35 +919,35 @@ func TestGetDirectionNetworkPolicy(t *testing.T) {
 	// L4+L7
 	xds := testXdsServer(t)
 	selectors := testSelectorCache.GetSelectorSnapshot()
-	obtained := xds.getDirectionNetworkPolicy(ep, nil, selectors, &L4Policy1.Ingress, true, false, false, "ingress", "")
+	obtained := GetDirectionNetworkPolicy(ep, nil, selectors, &L4Policy1.Ingress, true, false, false, "ingress", "", xds.logger, xds.l7RulesTranslator)
 	require.Equal(t, ExpectedPerPortPolicies12, obtained)
 
 	// L4+L7 with header mods
-	obtained = xds.getDirectionNetworkPolicy(ep, nil, selectors, &L4HeaderMatchPolicy1.Ingress, true, false, false, "ingress", "")
+	obtained = GetDirectionNetworkPolicy(ep, nil, selectors, &L4HeaderMatchPolicy1.Ingress, true, false, false, "ingress", "", xds.logger, xds.l7RulesTranslator)
 	require.Equal(t, ExpectedPerPortPolicies122HeaderMatch, obtained)
 
 	// L4+L7
-	obtained = xds.getDirectionNetworkPolicy(ep, nil, selectors, &L4Policy1.Egress, true, false, false, "egress", "")
+	obtained = GetDirectionNetworkPolicy(ep, nil, selectors, &L4Policy1.Egress, true, false, false, "egress", "", xds.logger, xds.l7RulesTranslator)
 	require.Equal(t, ExpectedPerPortPolicies1, obtained)
 
 	// L4+L7 with Deny L3
-	obtained = xds.getDirectionNetworkPolicy(ep, nil, selectors, &L4Deny2Policy1.Ingress, true, false, false, "ingress", "")
+	obtained = GetDirectionNetworkPolicy(ep, nil, selectors, &L4Deny2Policy1.Ingress, true, false, false, "ingress", "", xds.logger, xds.l7RulesTranslator)
 	require.Equal(t, ExpectedPerPortPolicies1Deny2, obtained)
 
 	// L4-only
-	obtained = xds.getDirectionNetworkPolicy(ep, nil, selectors, &L4Policy4.Ingress, true, false, false, "ingress", "")
+	obtained = GetDirectionNetworkPolicy(ep, nil, selectors, &L4Policy4.Ingress, true, false, false, "ingress", "", xds.logger, xds.l7RulesTranslator)
 	require.Equal(t, ExpectedPerPortPolicies, obtained)
 
 	// L4-only
-	obtained = xds.getDirectionNetworkPolicy(ep, nil, selectors, &L4Policy5.Ingress, true, false, false, "ingress", "")
+	obtained = GetDirectionNetworkPolicy(ep, nil, selectors, &L4Policy5.Ingress, true, false, false, "ingress", "", xds.logger, xds.l7RulesTranslator)
 	require.Equal(t, ExpectedPerPortPoliciesWildcard, obtained)
 
 	// L4-only with SNI
-	obtained = xds.getDirectionNetworkPolicy(ep, nil, selectors, &L4SNIPolicy.Ingress, true, false, false, "ingress", "")
+	obtained = GetDirectionNetworkPolicy(ep, nil, selectors, &L4SNIPolicy.Ingress, true, false, false, "ingress", "", xds.logger, xds.l7RulesTranslator)
 	require.Equal(t, ExpectedPerPortPoliciesSNI, obtained)
 
 	// with pass verdict
-	obtained = xds.getDirectionNetworkPolicy(ep, nil, selectors, &L4PassPolicy.Ingress, true, false, false, "ingress", "")
+	obtained = GetDirectionNetworkPolicy(ep, nil, selectors, &L4PassPolicy.Ingress, true, false, false, "ingress", "", xds.logger, xds.l7RulesTranslator)
 	require.Equal(t, ExpectedPerPortPoliciesPass, obtained)
 
 }
@@ -971,7 +981,7 @@ func TestGetDirectionNetworkPolicyCoalescesFiltersByResolvedPort(t *testing.T) {
 			}),
 		}
 
-		obtained := xds.getDirectionNetworkPolicy(ep, egressPortFunc, selectors, l4DirectionPolicy, true, false, false, "egress", "")
+		obtained := GetDirectionNetworkPolicy(ep, egressPortFunc, selectors, l4DirectionPolicy, true, false, false, "egress", "", xds.logger, xds.l7RulesTranslator)
 		require.Len(t, obtained, 1)
 		require.Equal(t, uint32(80), obtained[0].Port)
 		require.Equal(t, envoy_config_core.SocketAddress_TCP, obtained[0].Protocol)
@@ -1008,7 +1018,7 @@ func TestGetDirectionNetworkPolicyCoalescesFiltersByResolvedPort(t *testing.T) {
 			}),
 		}
 
-		obtained := xds.getDirectionNetworkPolicy(ep, egressPortFunc, selectors, l4DirectionPolicy, true, false, false, "egress", "")
+		obtained := GetDirectionNetworkPolicy(ep, egressPortFunc, selectors, l4DirectionPolicy, true, false, false, "egress", "", xds.logger, xds.l7RulesTranslator)
 		require.Equal(t, []*cilium.PortNetworkPolicy{{
 			Port:     80,
 			Protocol: envoy_config_core.SocketAddress_TCP,
@@ -1045,7 +1055,7 @@ func TestGetDirectionNetworkPolicyCoalescesFiltersByResolvedPort(t *testing.T) {
 			}),
 		}
 
-		obtained := xds.getDirectionNetworkPolicy(ep, egressPortFunc, selectors, l4DirectionPolicy, true, false, false, "egress", "")
+		obtained := GetDirectionNetworkPolicy(ep, egressPortFunc, selectors, l4DirectionPolicy, true, false, false, "egress", "", xds.logger, xds.l7RulesTranslator)
 		require.Equal(t, []*cilium.PortNetworkPolicy{{
 			Port:     80,
 			Protocol: envoy_config_core.SocketAddress_TCP,
@@ -1078,7 +1088,7 @@ func TestGetDirectionNetworkPolicyEgressNamedPorts(t *testing.T) {
 			}),
 		}
 
-		obtained := xds.getDirectionNetworkPolicy(ep, egressPortFunc, selectors, l4DirectionPolicy, true, false, false, "egress", "")
+		obtained := GetDirectionNetworkPolicy(ep, egressPortFunc, selectors, l4DirectionPolicy, true, false, false, "egress", "", xds.logger, xds.l7RulesTranslator)
 		require.Equal(t, []*cilium.PortNetworkPolicy{{
 			Port:     8080,
 			Protocol: envoy_config_core.SocketAddress_TCP,
@@ -1116,7 +1126,7 @@ func TestGetDirectionNetworkPolicyEgressNamedPorts(t *testing.T) {
 			}),
 		}
 
-		obtained := xds.getDirectionNetworkPolicy(ep, egressPortFunc, selectors, l4DirectionPolicy, true, false, false, "egress", "")
+		obtained := GetDirectionNetworkPolicy(ep, egressPortFunc, selectors, l4DirectionPolicy, true, false, false, "egress", "", xds.logger, xds.l7RulesTranslator)
 		require.Equal(t, []*cilium.PortNetworkPolicy{{
 			Port:     8080,
 			Protocol: envoy_config_core.SocketAddress_TCP,
@@ -1153,7 +1163,7 @@ func TestGetDirectionNetworkPolicyEgressNamedPorts(t *testing.T) {
 			}),
 		}
 
-		obtained := xds.getDirectionNetworkPolicy(ep, egressPortFunc, selectors, l4DirectionPolicy, true, false, false, "egress", "")
+		obtained := GetDirectionNetworkPolicy(ep, egressPortFunc, selectors, l4DirectionPolicy, true, false, false, "egress", "", xds.logger, xds.l7RulesTranslator)
 		require.Equal(t, []*cilium.PortNetworkPolicy{{
 			Port:     8080,
 			Protocol: envoy_config_core.SocketAddress_TCP,
@@ -1196,7 +1206,7 @@ func TestGetDirectionNetworkPolicyWildcardPass(t *testing.T) {
 			},
 		}), []types.Priority{0, 0x100})
 
-		obtained := xds.getDirectionNetworkPolicy(ep, nil, selectors, l4DirectionPolicy, true, false, false, "ingress", "")
+		obtained := GetDirectionNetworkPolicy(ep, nil, selectors, l4DirectionPolicy, true, false, false, "ingress", "", xds.logger, xds.l7RulesTranslator)
 		require.Equal(t, []*cilium.PortNetworkPolicy{{
 			Port:     0,
 			Protocol: envoy_config_core.SocketAddress_TCP,
@@ -1245,7 +1255,7 @@ func TestGetDirectionNetworkPolicyWildcardPass(t *testing.T) {
 			},
 		}), []types.Priority{0})
 
-		obtained := xds.getDirectionNetworkPolicy(ep, nil, selectors, l4DirectionPolicy, true, false, false, "ingress", "")
+		obtained := GetDirectionNetworkPolicy(ep, nil, selectors, l4DirectionPolicy, true, false, false, "ingress", "", xds.logger, xds.l7RulesTranslator)
 		require.Equal(t, []*cilium.PortNetworkPolicy{{
 			Port:     0,
 			Protocol: envoy_config_core.SocketAddress_TCP,
@@ -1298,7 +1308,7 @@ func TestGetDirectionNetworkPolicyWildcardPass(t *testing.T) {
 			},
 		}), []types.Priority{0})
 
-		obtained := xds.getDirectionNetworkPolicy(ep, nil, selectors, l4DirectionPolicy, true, false, false, "ingress", "")
+		obtained := GetDirectionNetworkPolicy(ep, nil, selectors, l4DirectionPolicy, true, false, false, "ingress", "", xds.logger, xds.l7RulesTranslator)
 		require.Equal(t, []*cilium.PortNetworkPolicy{{
 			Port:     0,
 			Protocol: envoy_config_core.SocketAddress_TCP,
@@ -1432,7 +1442,7 @@ func TestGetDirectionNetworkPolicyWildcardRedirect(t *testing.T) {
 				}),
 			}
 
-			obtained := xds.getDirectionNetworkPolicy(redirectEP, nil, selectors, l4DirectionPolicy, true, false, false, "ingress", "")
+			obtained := GetDirectionNetworkPolicy(redirectEP, nil, selectors, l4DirectionPolicy, true, false, false, "ingress", "", xds.logger, xds.l7RulesTranslator)
 			require.Equal(t, tc.expected, obtained)
 		})
 	}
@@ -1506,7 +1516,7 @@ func TestCNPWildcardPortListenerRedirectToEnvoy(t *testing.T) {
 		epp.Detach(logger)
 	})
 
-	obtained := xds.getDirectionNetworkPolicy(
+	obtained := GetDirectionNetworkPolicy(
 		redirectEP,
 		nil,
 		epp.GetPolicySelectors(),
@@ -1516,6 +1526,8 @@ func TestCNPWildcardPortListenerRedirectToEnvoy(t *testing.T) {
 		false,
 		"egress",
 		"",
+		xds.logger,
+		xds.l7RulesTranslator,
 	)
 
 	require.Equal(t, []*cilium.PortNetworkPolicy{{
@@ -1531,7 +1543,7 @@ func TestCNPWildcardPortListenerRedirectToEnvoy(t *testing.T) {
 func TestGetNetworkPolicy(t *testing.T) {
 	xds := testXdsServer(t)
 	selectors := testSelectorCache.GetSelectorSnapshot()
-	obtained := xds.getNetworkPolicy(ep, nil, selectors, []string{IPv4Addr}, L4Policy1, true, true, false, false, "")
+	obtained := GetNetworkPolicy(ep, nil, selectors, []string{IPv4Addr}, L4Policy1, true, true, false, false, "", xds.logger, xds.l7RulesTranslator)
 	expected := &cilium.NetworkPolicy{
 		EndpointIps:            []string{IPv4Addr},
 		EndpointId:             uint64(ep.GetID()),
@@ -1544,7 +1556,7 @@ func TestGetNetworkPolicy(t *testing.T) {
 func TestGetNetworkPolicyWildcard(t *testing.T) {
 	xds := testXdsServer(t)
 	selectors := testSelectorCache.GetSelectorSnapshot()
-	obtained := xds.getNetworkPolicy(ep, nil, selectors, []string{IPv4Addr}, L4Policy2, true, true, false, false, "")
+	obtained := GetNetworkPolicy(ep, nil, selectors, []string{IPv4Addr}, L4Policy2, true, true, false, false, "", xds.logger, xds.l7RulesTranslator)
 	expected := &cilium.NetworkPolicy{
 		EndpointIps:            []string{IPv4Addr},
 		EndpointId:             uint64(ep.GetID()),
@@ -1557,7 +1569,7 @@ func TestGetNetworkPolicyWildcard(t *testing.T) {
 func TestGetNetworkPolicyDeny(t *testing.T) {
 	xds := testXdsServer(t)
 	selectors := testSelectorCache.GetSelectorSnapshot()
-	obtained := xds.getNetworkPolicy(ep, nil, selectors, []string{IPv4Addr}, L4Policy1RequiresV2, true, true, false, false, "")
+	obtained := GetNetworkPolicy(ep, nil, selectors, []string{IPv4Addr}, L4Policy1RequiresV2, true, true, false, false, "", xds.logger, xds.l7RulesTranslator)
 	expected := &cilium.NetworkPolicy{
 		EndpointIps:            []string{IPv4Addr},
 		EndpointId:             uint64(ep.GetID()),
@@ -1570,7 +1582,7 @@ func TestGetNetworkPolicyDeny(t *testing.T) {
 func TestGetNetworkPolicyWildcardDeny(t *testing.T) {
 	xds := testXdsServer(t)
 	selectors := testSelectorCache.GetSelectorSnapshot()
-	obtained := xds.getNetworkPolicy(ep, nil, selectors, []string{IPv4Addr}, L4Policy1RequiresV2, true, true, false, false, "")
+	obtained := GetNetworkPolicy(ep, nil, selectors, []string{IPv4Addr}, L4Policy1RequiresV2, true, true, false, false, "", xds.logger, xds.l7RulesTranslator)
 	expected := &cilium.NetworkPolicy{
 		EndpointIps:            []string{IPv4Addr},
 		EndpointId:             uint64(ep.GetID()),
@@ -1583,7 +1595,7 @@ func TestGetNetworkPolicyWildcardDeny(t *testing.T) {
 func TestGetNetworkPolicyNil(t *testing.T) {
 	xds := testXdsServer(t)
 	selectors := testSelectorCache.GetSelectorSnapshot()
-	obtained := xds.getNetworkPolicy(ep, nil, selectors, []string{IPv4Addr}, nil, true, true, false, false, "")
+	obtained := GetNetworkPolicy(ep, nil, selectors, []string{IPv4Addr}, nil, true, true, false, false, "", xds.logger, xds.l7RulesTranslator)
 	expected := &cilium.NetworkPolicy{
 		EndpointIps:            []string{IPv4Addr},
 		EndpointId:             uint64(ep.GetID()),
@@ -1596,7 +1608,7 @@ func TestGetNetworkPolicyNil(t *testing.T) {
 func TestGetNetworkPolicyIngressNotEnforced(t *testing.T) {
 	xds := testXdsServer(t)
 	selectors := testSelectorCache.GetSelectorSnapshot()
-	obtained := xds.getNetworkPolicy(ep, nil, selectors, []string{IPv4Addr}, L4Policy2, false, true, false, false, "")
+	obtained := GetNetworkPolicy(ep, nil, selectors, []string{IPv4Addr}, L4Policy2, false, true, false, false, "", xds.logger, xds.l7RulesTranslator)
 	expected := &cilium.NetworkPolicy{
 		EndpointIps:            []string{IPv4Addr},
 		EndpointId:             uint64(ep.GetID()),
@@ -1609,7 +1621,7 @@ func TestGetNetworkPolicyIngressNotEnforced(t *testing.T) {
 func TestGetNetworkPolicyEgressNotEnforced(t *testing.T) {
 	xds := testXdsServer(t)
 	selectors := testSelectorCache.GetSelectorSnapshot()
-	obtained := xds.getNetworkPolicy(ep, nil, selectors, []string{IPv4Addr}, L4Policy1RequiresV2, true, false, false, false, "")
+	obtained := GetNetworkPolicy(ep, nil, selectors, []string{IPv4Addr}, L4Policy1RequiresV2, true, false, false, false, "", xds.logger, xds.l7RulesTranslator)
 	expected := &cilium.NetworkPolicy{
 		EndpointIps:            []string{IPv4Addr},
 		EndpointId:             uint64(ep.GetID()),
@@ -2301,7 +2313,7 @@ func TestGetNetworkPolicyTLSInterception(t *testing.T) {
 		t.Run(tt.name, func(t *testing.T) {
 			xds := testXdsServer(t)
 			selectors := testSelectorCache.GetSelectorSnapshot()
-			obtained := xds.getNetworkPolicy(ep, nil, selectors, []string{IPv4Addr}, tt.args.inputPolicy, true, true, tt.args.useFullTLSContext, tt.args.useSDS, tt.args.policySecretsNamespace)
+			obtained := GetNetworkPolicy(ep, nil, selectors, []string{IPv4Addr}, tt.args.inputPolicy, true, true, tt.args.useFullTLSContext, tt.args.useSDS, tt.args.policySecretsNamespace, xds.logger, xds.l7RulesTranslator)
 			expected := &cilium.NetworkPolicy{
 				EndpointIps:            []string{IPv4Addr},
 				EndpointId:             uint64(ep.GetID()),
@@ -2313,7 +2325,7 @@ func TestGetNetworkPolicyTLSInterception(t *testing.T) {
 	}
 }
 
-func Test_getPublicListenerAddress(t *testing.T) {
+func Test_GetPublicListenerAddress(t *testing.T) {
 	type args struct {
 		port uint16
 		ipv4 bool
@@ -2379,13 +2391,13 @@ func Test_getPublicListenerAddress(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			got := getPublicListenerAddress(tt.args.port, tt.args.ipv4, tt.args.ipv6)
+			got := GetPublicListenerAddress(tt.args.port, tt.args.ipv4, tt.args.ipv6)
 			assert.Equal(t, tt.want, got)
 		})
 	}
 }
 
-func Test_getLocalListenerAddresses(t *testing.T) {
+func Test_GetLocalListenerAddresses(t *testing.T) {
 	v4Local := &envoy_config_core.Address_SocketAddress{
 		SocketAddress: &envoy_config_core.SocketAddress{
 			Protocol:      envoy_config_core.SocketAddress_TCP,
@@ -2488,7 +2500,7 @@ func TestUpdateNetworkPolicyRevertKeepsLocalEndpointStoreAfterStaleDuplicateRemo
 	repo, localIdentity, currentEPP := newTestEndpointPolicy(t, currentEP)
 	xds := newTestXDSServer(t)
 
-	err, revert, finalize := xds.UpdateNetworkPolicy(currentEP, currentEPP, nil)
+	err, revert, finalize := xds.UpdateNetworkPolicy(t.Context(), currentEP, currentEPP, nil)
 	require.NoError(t, err)
 	require.NotNil(t, revert)
 	require.NotNil(t, finalize)
@@ -2509,7 +2521,7 @@ func TestUpdateNetworkPolicyRevertKeepsLocalEndpointStoreAfterStaleDuplicateRemo
 	require.Equal(t, staleEP.GetID(), localEP.GetID())
 
 	refreshedCurrentEPP := distillEndpointPolicy(t, repo, localIdentity, currentEP)
-	err, revert, finalize = xds.UpdateNetworkPolicy(currentEP, refreshedCurrentEPP, nil)
+	err, revert, finalize = xds.UpdateNetworkPolicy(t.Context(), currentEP, refreshedCurrentEPP, nil)
 	require.NoError(t, err)
 	require.NotNil(t, revert)
 	require.NotNil(t, finalize)
@@ -2536,6 +2548,37 @@ func TestUpdateNetworkPolicyRevertKeepsLocalEndpointStoreAfterStaleDuplicateRemo
 	stalePolicy, err := xds.networkPolicyCache.Lookup(NetworkPolicyTypeURL, staleResourceName)
 	require.NoError(t, err)
 	require.Nil(t, stalePolicy)
+}
+
+func TestUpdateNetworkPolicyLegacyACKUsesNodeIP(t *testing.T) {
+	currentEP := &listenerProxyUpdaterMock{ProxyUpdaterMock: &test.ProxyUpdaterMock{
+		Id:   2766,
+		Ipv4: "10.0.0.159",
+		Ipv6: "fd00:10:244::ab0e",
+	}}
+	_, _, currentEPP := newTestEndpointPolicy(t, currentEP)
+	xdsServer := newTestXDSServer(t)
+	_ = xdsServer.npdsListeners.Add("test-proxy-listener")
+
+	ctx, cancel := context.WithTimeout(context.Background(), time.Second)
+	defer cancel()
+	wg := completion.NewWaitGroup(ctx)
+
+	err, revert, finalize := xdsServer.UpdateNetworkPolicy(ctx, currentEP, currentEPP, wg)
+	require.NoError(t, err)
+	require.NotNil(t, revert)
+	require.NotNil(t, finalize)
+
+	acker, ok := xdsServer.networkPolicyMutator.(*xds.AckingResourceMutatorWrapper)
+	require.True(t, ok)
+	resourceName := strconv.FormatUint(currentEP.GetID(), 10)
+	resources, err := xdsServer.networkPolicyCache.GetResources(NetworkPolicyTypeURL, 0, "127.0.0.1", nil)
+	require.NoError(t, err)
+	acker.HandleResourceVersionAck(resources.Version, resources.Version, "127.0.0.1", []string{resourceName}, NetworkPolicyTypeURL, "")
+
+	require.NoError(t, wg.Wait())
+	finalize()
+	require.NoError(t, revert())
 }
 
 func newTestEndpointPolicy(t *testing.T, ep *listenerProxyUpdaterMock) (*policy.Repository, *identity.Identity, *policy.EndpointPolicy) {
@@ -2592,7 +2635,7 @@ func newTestXDSServer(t *testing.T) *xdsServer {
 		nil,
 		nil,
 		newLocalEndpointStore(),
-		xdsServerConfig{},
+		xdsServerConfig{metrics: noopXDSMetrics{}},
 		certificatemanager.NewMockSecretManagerInline(),
 	)
 	xds.l7RulesTranslator = envoypolicy.NewEnvoyL7RulesTranslator(logger, nil)
